+++ x/gov/types/v1/genesis.go
@@ -38,7 +38,7 @@ func (data GenesisState) Empty() bool {
 // ValidateGenesis checks if parameters are within valid ranges
 func ValidateGenesis(data *GenesisState) error {
 	if data.StartingProposalId == 0 {
-		return errors.New("Starting proposal id must be greater than 0")
+		return errors.New("starting proposal id must be greater than 0")
 	}
 
 	if err := validateTallyParams(*data.TallyParams); err != nil {